ci: ASan/UBSan leg + drop unused CMakePresets
Add an Ubuntu AddressSanitizer + UBSan job to build.yml: RelWithDebInfo
for line numbers, leaks off (focus on corruption/UB, not exit-time test
leaks), UBSan halt_on_error so a violation fails the build. Exclude the
alignment sub-check -- the SSE bit-transpose (block.hpp sse_trans) does
deliberate unaligned 2-byte stores, well-defined on x86-64/arm64 and only
reachable from sub-16 test dimensions; real callers transpose 128-row
(aligned) blocks. Every other UBSan check stays on.

The sanitizer surfaced a genuine signed-integer-overflow in
test_runtime_int's reference product (-1234567 * 7654321 overflows int32);
compute it in unsigned so the reference matches the runtime Int's
2's-complement wrap without UB.

Delete CMakePresets.json: unreferenced by CI, README, or any documented
build -- the repo configures via raw -DCMAKE_BUILD_TYPE everywhere.

+++ test/test_runtime_int.cpp
@@ -60,7 +60,9 @@ int main() {
   const int32_t SA = -1234567, SB = 7654321;
   DI sa = DI::constant(cx, 32, SA), sb = DI::constant(cx, 32, SB);
   chk("signed add", rds(sa + sb) == (int32_t)(SA + SB));
-  chk("signed mul", rds(sa * sb) == (int32_t)(SA * SB));
+  // Reference wraps mod 2^32 (matches the runtime Int's 2's-complement mul);
+  // compute it in unsigned to keep the reference itself free of signed overflow.
+  chk("signed mul", rds(sa * sb) == (int32_t)((uint32_t)SA * (uint32_t)SB));
   chk("signed div", rds(sb / sa) == (SB / SA));
   chk("signed mod", rds(sb % sa) == (SB % SA));
   chk("signed lt", ((sa < sb).w & 1) == 1);
